completed security exercise allowing only logged in users to create posts and assigns logged in users to created post

+++ src/main/java/com/codeup/springblog/controllers/MathController.java
@@ -8,31 +8,31 @@ import org.springframework.web.bind.annotation.*;
 public class MathController {
     @RequestMapping(path = "/add100To/{number}", method = RequestMethod.GET)
     @ResponseBody
-    public String addOneHundred(@PathVariable int number){
+    public String addOneHundred(@PathVariable double number){
         return number + " + 100 = " + (number + 100);
     }
 
     @RequestMapping(path = "/add/{number1}/and/{number2}", method = RequestMethod.GET)
     @ResponseBody
-    public Integer addTwoNumbers(@PathVariable int number1, @PathVariable int number2){
+    public double addTwoNumbers(@PathVariable double number1, @PathVariable double number2){
         return number1 + number2;
     }
 
     @RequestMapping(path = "/subtract/{number1}/and/{number2}", method = RequestMethod.GET)
     @ResponseBody
-    public Integer subTwoNumbers(@PathVariable int number1, @PathVariable int number2){
+    public double subTwoNumbers(@PathVariable double number1, @PathVariable double number2){
         return number1 - number2;
     }
 
     @RequestMapping(path = "/multiply/{number1}/and/{number2}", method = RequestMethod.GET)
     @ResponseBody
-    public Integer multiplyTwoNumbers(@PathVariable int number1, @PathVariable int number2){
+    public double multiplyTwoNumbers(@PathVariable double number1, @PathVariable double number2){
         return number1 * number2;
     }
 
     @RequestMapping(path = "/divide/{number1}/and/{number2}", method = RequestMethod.GET)
     @ResponseBody
-    public Integer divTwoNumbers(@PathVariable int number1, @PathVariable int number2){
+    public double divTwoNumbers(@PathVariable double number1, @PathVariable double number2){
         return number1 / number2;
     }
